+++ Imperva/tests/test_file_downloader.py
@@ -15,10 +15,10 @@ def test_file_downloader_200_with_proxy(config):
 
     with patch("urllib3.ProxyManager") as m:
         mock_http = m.return_value
-        mock_http.request.return_value = Mock(status=200, data="some data")
+        mock_http.request.return_value = Mock(status=200, data=b"some data")
 
         result = fd.request_file_content(config.BASE_URL, timeout=1)
-        assert result == "some data"
+        assert result == b"some data"
 
         mock_http.request.assert_called_once_with(
             "GET", config.BASE_URL, headers={"authorization": "Basic bG9yZW06aXBzdW0="}
@@ -30,11 +30,11 @@ def test_file_downloader_404(config):
 
     with patch("urllib3.PoolManager") as m:
         mock_http = m.return_value
-        mock_http.request.return_value = Mock(status=404, data="not found")
+        mock_http.request.return_value = Mock(status=404, data=b"not found")
 
         result = fd.request_file_content(config.BASE_URL, timeout=1)
 
-        assert result == ""
+        assert result == b""
 
         mock_http.request.assert_called_once_with(
             "GET", config.BASE_URL, headers={"authorization": "Basic bG9yZW06aXBzdW0="}
@@ -79,7 +79,7 @@ def test_file_downloader_other(config):
         mock_http.request.return_value = Mock(status=500)
 
         result = fd.request_file_content(config.BASE_URL, timeout=1)
-        assert result == ""
+        assert result == b""
 
         mock_http.request.assert_called_once_with(
             "GET", config.BASE_URL, headers={"authorization": "Basic bG9yZW06aXBzdW0="}