+++ PagerDuty/pagerduty/helpers.py
@@ -6,7 +6,7 @@ Various helping functions
 # third parties
 import requests
 from requests.adapters import HTTPAdapter
-from requests.packages.urllib3.util.retry import Retry
+from urllib3.util.retry import Retry
 
 
 def requests_retry_session(