+++ Azure/CHANGELOG.md
@@ -6,3 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 2023-10-24 - 2.1.1
+
+### Fixed
+
+- Added Beta Flag for Azure Blob Storage
+
+## 2023-10-20 - 2.0.1
+
+### Fixed
+
+- Fix the import of connector
+
+## 2023-10-02 - 2.0.0
+
+### Added
+
+- Add new connector to fetch events from Azure Blob storage.
+
+### Changed
+
+- Refactor the module